Move to notebook_base_image.py
