+++ modal_global_objects/images/base_images.py
@@ -9,6 +9,8 @@ from rich.table import Table
 import modal
 from modal.image import SUPPORTED_PYTHON_SERIES, ImageBuilderVersion
 
+from .notebook_base_image import notebook_base_image
+
 
 def dummy():
     pass
@@ -16,7 +18,13 @@ def dummy():
 
 if __name__ == "__main__":
     _, name = sys.argv
-    constructor = getattr(modal.Image, name)
+
+    if name in ("debian_slim", "micromamba"):
+        constructor = getattr(modal.Image, name)
+    elif name == "notebook":
+        constructor = notebook_base_image
+    else:
+        raise ValueError(f"Unknown base image type: {name}")
 
     builder_version = os.environ.get("MODAL_IMAGE_BUILDER_VERSION")
     assert builder_version, "Script requires MODAL_IMAGE_BUILDER_VERSION environment variable"
